Feat/remember failed lyrics imports (#20)
* feat: retry to import lyrics only 3 times

After failing to acquire the lyrics 3 times the track is marked as failed. This should prevent the app from piling up a lot of tracks during the import where no lyrics can ever be found. The sidebar now contains an extra link to a list of those tracks. These tracks can be reviewed and the lyrics can be entered manually.

* fix: print correct error message during lyrics import

* feat: add title to tracks page